Reorganize cleanup service
Reorganize the cleanup service to make the code easier to maintain and
to provide a more logical file layout. Some of the changes:

- Extract `performCleanup` from `StartDbCleanupService`
- `DeleteIncompleteLogs` -> `RemoveIncompleteLogs`
- Move functions from `dbcleanupservice.go` and
  `dbcleanupservice_test.go` to more appropriate files.
