+++ internal/database/dbcleanupservice/BUILD.bazel
@@ -6,15 +6,15 @@ go_library(
         "batcher.go",
         "compact_logs.go",
         "dbcleanupservice.go",
-        "delete_incomplete_logs.go",
         "lockinvocations.go",
         "next_slice.go",
+        "remove_builds.go",
         "remove_files.go",
+        "remove_inactive_users.go",
         "remove_old_invocations.go",
         "remove_orphaned_test_targets.go",
         "remove_target_kind_mappings.go",
         "remove_unused_targets.go",
-        "removeinactiveusers.go",
     ],
     importpath = "github.com/buildbarn/bb-portal/internal/database/dbcleanupservice",
     visibility = ["//:__subpackages__"],
@@ -48,15 +48,17 @@ go_test(
     name = "dbcleanupservice_test",
     srcs = [
         "batcher_test.go",
+        "compact_logs_test.go",
         "dbcleanupservice_test.go",
         "lockinvocations_test.go",
         "next_slice_test.go",
+        "remove_builds_test.go",
         "remove_files_test.go",
+        "remove_inactive_users_test.go",
         "remove_old_invocations_test.go",
         "remove_orphaned_test_targets_test.go",
         "remove_target_kind_mappings_test.go",
         "remove_unused_targets_test.go",
-        "removeinactiveusers_test.go",
     ],
     deps = [
         ":dbcleanupservice",
